configure apt to allow easyle manipulation and minimize security
* we will use only debian repos so then to allow easy build
we need to minimize security, no check of certificates neither
check of sources, of course final build will revert this!

* `Acquire::Check-Valid-Until "0";` bypass release file, cos
sometimes this file is not in sync with new packages on repos or
new debian versions
* `Acquire::Languages none;` no download the translations of
the package list and descriptions, it will download all the languajes
and will consume more bandwith
* `Acquire::IndexTargets::deb::Contents-deb::DefaultEnabled false;`
will skip updating contents of package description.. on the command
update only, this will reduce consuption and bandwitch usage
* `Apt::AutoRemove::InstallRecommends "1";` remove the recommends on
the removal of a package that install those ones
* `APT::Install-Recommends "1";` set to true/1 to always install
apart of the desired package any recommendation of such package, this
has a undesired effect that install many things that not always are
enabled or well integrated, but for easy of use for blind people is
very necesary here to be enabled!
* `APT::Get::AllowUnauthenticated "1";` allows repositories that are
not configured very well for high security
* `APT::Get::AutomaticRemove "0";` disable removal of packages on
apt operations that are not used anymore
* `APT::Get::HideAutoRemove "1";` dont allow or notify about the
auto removal operations on any apt installation or removal
* `APT::Get::AutoRemove "0";` same as previous
* `APT::Force-LoopBreak "true";` allow downgrade/upgrade of packages
that are essentials or may change important parts of the system,
necesary to allow custom configurations or mix repos, this will be
reverted inside system for future.
* `DPkg::Options {"--force-confnew";}` always use new configurations
of new packages or upgrades. dischard older configs, we prefers this
cos not always migrations works and new configus just breaks things
* `APT::Periodic::Update-Package-Lists "0";` dont auto upgrade the list
of packages in the installed system, we are not mocosoft
* `APT::Periodic::Download-Upgradeable-Packages "0";` same as previous
but dont allow upgrade automatically, only if user wants